Implement roll-table and probability-table syntax
A roll table is a two-column table:

    !!!roll-table
    <title>
    <n>,<markdown description>
    ...

A probability table is an at-least-two column table:

    !!!probability-table
    <n>,<probability>,...

Which is rendered with a bar chart after each column of probabilities.
The bars are scaled to the largest integer multiple which fits, across
the whole table.  This is because a bar of, say, 10% width in a fairly
narrow column isn't very easy to read.