Add an optional "header image" to posts
If a file `static/headers/{slug}.[png,jpg]` exists, it'll be displayed
at the top of the post.  These imaghes aren't necessarily the same as
the Twitter cards, because the Twitter cards might duplicate the title
or some image in the post, and so just look bad or be confusing.